Basic mesh loading using assimp! For now the ship is replaced by the mesh of a bunny. Don't ask.

+++ Ex3/src/HealthBar.cpp
@@ -27,7 +27,7 @@ _maxUnits(maxUnits), _currentUnits(maxUnits)
         1, 3, 2
     };
     
-    std::vector<GLubyte> indicesVec(indices, indices + (sizeof(indices) / sizeof(GLubyte)));
+    std::vector<GLuint> indicesVec(indices, indices + (sizeof(indices) / sizeof(GLuint)));
     _renderComponent->setIBO(indicesVec);
 }
